Timetable Changes
Data now goes into the correct field for the timetables - for this to
work admin team need to ensure that when a new lesson is added, all
students on that module have record created on the register. If it does
not appear as a lesson on the register for them, it will not appear on
the timetable.